requirements.txt
openai-whisper
nltk